Refactor proof artifact helpers into submodules
